+++ src/components/Navigation/AppBreadcrumbs/AppBreadcrumbs.types.ts
@@ -1,0 +1,1 @@
+export type PathSegment = { title: string; path: string };
