+++ src/index.css
@@ -51,7 +51,6 @@ tds-header {
   gap: 32px;
   max-width: 600px;
   justify-content: center;
-  /* margin: 0 auto; */
   padding: 128px 0;
 }
